完成了主页歌单页面，分享ui以完成

+++ app/src/main/java/com/example/musicplayerdome/main/bean/HighQualityPlayListBean.java
@@ -487,7 +487,7 @@ public class HighQualityPlayListBean {
             private int gender;
             private int city;
             private long birthday;
-            private int userId;
+            private long userId;
             private int userType;
             private String nickname;
             private String signature;
@@ -579,11 +579,11 @@ public class HighQualityPlayListBean {
                 this.birthday = birthday;
             }
 
-            public int getUserId() {
+            public long getUserId() {
                 return userId;
             }
 
-            public void setUserId(int userId) {
+            public void setUserId(long userId) {
                 this.userId = userId;
             }
 
@@ -773,7 +773,7 @@ public class HighQualityPlayListBean {
             private int gender;
             private int city;
             private long birthday;
-            private int userId;
+            private long userId;
             private int userType;
             private String nickname;
             private String signature;
@@ -865,11 +865,11 @@ public class HighQualityPlayListBean {
                 this.birthday = birthday;
             }
 
-            public int getUserId() {
+            public long getUserId() {
                 return userId;
             }
 
-            public void setUserId(int userId) {
+            public void setUserId(long userId) {
                 this.userId = userId;
             }
